requirements.txt
huggingface-hub
textblob
nltk
langchain
python-dotenv
chainlit
literalai
